+++ src/edu/washington/escience/myria/api/encoding/DupElimEncoding.java
@@ -3,12 +3,11 @@ package edu.washington.escience.myria.api.encoding;
 import edu.washington.escience.myria.api.MyriaApiException;
 import edu.washington.escience.myria.api.encoding.QueryConstruct.ConstructArgs;
 import edu.washington.escience.myria.operator.DupElim;
-import edu.washington.escience.myria.operator.StreamingStateWrapper;
 
-public class DupElimEncoding extends UnaryOperatorEncoding<StreamingStateWrapper> {
+public class DupElimEncoding extends UnaryOperatorEncoding<DupElim> {
 
   @Override
-  public StreamingStateWrapper construct(ConstructArgs args) throws MyriaApiException {
-    return new StreamingStateWrapper(null, new DupElim());
+  public DupElim construct(ConstructArgs args) throws MyriaApiException {
+    return new DupElim(null);
   }
 }
