+++ src/edu/washington/escience/myria/api/encoding/DupElimStateEncoding.java
@@ -1,11 +1,11 @@
 package edu.washington.escience.myria.api.encoding;
 
-import edu.washington.escience.myria.operator.DupElim;
+import edu.washington.escience.myria.operator.StatefulDupElim;
 
-public class DupElimStateEncoding extends StreamingStateEncoding<DupElim> {
+public class DupElimStateEncoding extends StreamingStateEncoding<StatefulDupElim> {
 
   @Override
-  public DupElim construct() {
-    return new DupElim();
+  public StatefulDupElim construct() {
+    return new StatefulDupElim();
   }
 }
