+++ src/edu/washington/escience/myria/operator/DupElim.java
@@ -1,87 +1,110 @@
 package edu.washington.escience.myria.operator;
 
+import java.nio.ByteBuffer;
 import java.util.BitSet;
-import java.util.List;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
-import com.gs.collections.api.block.procedure.primitive.IntProcedure;
-import com.gs.collections.impl.list.mutable.primitive.IntArrayList;
-import com.gs.collections.impl.map.mutable.primitive.IntObjectHashMap;
+import com.gs.collections.api.iterator.LongIterator;
+import com.gs.collections.impl.list.mutable.primitive.LongArrayList;
+import com.gs.collections.impl.map.mutable.primitive.LongLongHashMap;
+import com.gs.collections.impl.map.mutable.primitive.LongObjectHashMap;
 
+import edu.washington.escience.myria.DbException;
 import edu.washington.escience.myria.Schema;
-import edu.washington.escience.myria.column.Column;
-import edu.washington.escience.myria.storage.MutableTupleBuffer;
 import edu.washington.escience.myria.storage.TupleBatch;
-import edu.washington.escience.myria.storage.TupleUtils;
 import edu.washington.escience.myria.util.HashUtils;
 
 /**
- * Duplicate elimination. It adds newly meet unique tuples into a buffer so that the source TupleBatches are not
- * referenced. This implementation reduces memory consumption.
+ * Instead of storing unique tuples to deduplicate against,
+ * this implementation stores 128-bit hash codes of tuples
+ * (to make collisions statistically negligible).
  * */
-public final class DupElim extends StreamingState {
+public final class DupElim extends UnaryOperator {
 
   /** Required for Java serialization. */
   private static final long serialVersionUID = 1L;
+  /**
+   * We store this value instead of a valid hash code to indicate
+   * that a given initial hash code is mapped to multiple final hash codes.
+   */
+  private static final long COLLIDING_KEY = -1;
+  /**
+   * We return this value from getIfAbsent() to indicate absence,
+   * since 0 and -1 are already legitimate values.
+   */
+  private static final long ABSENT_VALUE = -2;
+  /** Stores 128-bit hash codes as map of first 64-bit word to second 64-bit word. */
+  private transient LongLongHashMap initialToFinalhashCodes;
+  /** Map of first 64-bit words to colliding second 64-bit words. */
+  private transient LongObjectHashMap<LongArrayList> collidingFinalHashCodes;
+  /** Cached byte buffer to avoid allocating a new buffer for each byte copy. */
+  private transient ByteBuffer byteCopyBuffer = ByteBuffer.allocate(Long.BYTES);
 
   /**
-   * The logger for this class.
+   * @param child the child
    * */
-  static final Logger LOGGER = LoggerFactory.getLogger(DupElim.class);
-
-  /**
-   * Indices to unique tuples.
-   * */
-  private transient IntObjectHashMap<IntArrayList> uniqueTupleIndices;
-
-  /**
-   * The buffer for storing unique tuples.
-   * */
-  private transient MutableTupleBuffer uniqueTuples = null;
-
-  @Override
-  public void cleanup() {
-    uniqueTuples = null;
-    uniqueTupleIndices = null;
+  public DupElim(final Operator child) {
+    super(child);
   }
 
+  @Override
+  protected void cleanup() throws DbException {}
+
   /**
-   * Do duplicate elimination for tb.
+   * Do duplicate elimination for the tb.
    *
-   * @param tb the TupleBatch for performing DupElim.
-   * @return the duplicate eliminated TB.
+   * @param tb the TB.
+   * @return a new TB with duplicates removed.
    * */
   protected TupleBatch doDupElim(final TupleBatch tb) {
     final int numTuples = tb.numTuples();
-    /* if tb is empty, directly return. */
     if (numTuples <= 0) {
       return tb;
     }
-    checkUniqueness.inputTB = tb;
-    List<? extends Column<?>> columns = tb.getDataColumns();
     final BitSet toRemove = new BitSet(numTuples);
     for (int i = 0; i < numTuples; ++i) {
-      final int nextIndex = uniqueTuples.numTuples();
-      final int cntHashCode = HashUtils.hashRow(tb, i);
-      IntArrayList tupleIndexList = uniqueTupleIndices.get(cntHashCode);
-      checkUniqueness.row = i;
-      checkUniqueness.unique = true;
-      if (tupleIndexList == null) {
-        tupleIndexList = new IntArrayList(1);
-        tupleIndexList.add(nextIndex);
-        uniqueTupleIndices.put(cntHashCode, tupleIndexList);
-      } else {
-        tupleIndexList.forEach(checkUniqueness);
-      }
-      if (checkUniqueness.unique) {
-        for (int j = 0; j < tb.numColumns(); ++j) {
-          uniqueTuples.put(j, columns.get(j), i);
+      final byte[] hashCodeBytes = HashUtils.hashRowBytes(tb, i);
+      Preconditions.checkArgument(
+          hashCodeBytes.length == 16,
+          "Expected 16 bytes in hash code, found %d",
+          hashCodeBytes.length);
+      byteCopyBuffer.clear();
+      byteCopyBuffer.put(hashCodeBytes, 0, 8);
+      byteCopyBuffer.flip();
+      long hashCode1 = byteCopyBuffer.getLong();
+      byteCopyBuffer.clear();
+      byteCopyBuffer.put(hashCodeBytes, 8, 8);
+      byteCopyBuffer.flip();
+      long hashCode2 = byteCopyBuffer.getLong();
+      long hashCodeValue = initialToFinalhashCodes.getIfAbsent(hashCode1, ABSENT_VALUE);
+      if (hashCodeValue == ABSENT_VALUE) {
+        initialToFinalhashCodes.put(hashCode1, hashCode2);
+      } else if (hashCodeValue == COLLIDING_KEY) {
+        LongArrayList collidingHashCodes = collidingFinalHashCodes.get(hashCode1);
+        Preconditions.checkNotNull(collidingHashCodes);
+        Preconditions.checkState(collidingHashCodes.size() > 1);
+        LongIterator iter = collidingHashCodes.longIterator();
+        boolean found = false;
+        while (iter.hasNext()) {
+          long hc = iter.next();
+          if (hc == hashCode2) {
+            // duplicate found
+            toRemove.set(i);
+            found = true;
+            break;
+          }
         }
-        tupleIndexList.add(nextIndex);
+        if (!found) {
+          collidingHashCodes.add(hashCode2);
+        }
+      } else if (hashCodeValue != hashCode2) {
+        LongArrayList collidingHashCodes = LongArrayList.newListWith(hashCodeValue, hashCode2);
+        Preconditions.checkState(!collidingFinalHashCodes.containsKey(hashCode1));
+        collidingFinalHashCodes.put(hashCode1, collidingHashCodes);
+        initialToFinalhashCodes.put(hashCode1, COLLIDING_KEY);
       } else {
+        // duplicate found
         toRemove.set(i);
       }
     }
@@ -89,71 +112,32 @@ public final class DupElim extends StreamingState {
   }
 
   @Override
-  public Schema getSchema() {
-    return getOp().getInputSchema();
-  }
+  public TupleBatch fetchNextReady() throws DbException {
 
-  @Override
-  public void init(final ImmutableMap<String, Object> execEnvVars) {
-    uniqueTupleIndices = new IntObjectHashMap<>();
-    uniqueTuples = new MutableTupleBuffer(getSchema());
-    checkUniqueness = new CheckUniquenessProcedure();
-  }
-
-  @Override
-  public TupleBatch update(final TupleBatch tb) {
-    TupleBatch newtb = doDupElim(tb);
-    if (newtb.numTuples() > 0 || tb.isEOI()) {
-      return newtb;
+    TupleBatch tb = null;
+    tb = getChild().nextReady();
+    while (tb != null) {
+      tb = doDupElim(tb);
+      if (tb.numTuples() > 0) {
+        return tb;
+      }
+      tb = getChild().nextReady();
     }
     return null;
   }
 
   @Override
-  public List<TupleBatch> exportState() {
-    return uniqueTuples.getAll();
+  public Schema generateSchema() {
+    Operator child = getChild();
+    if (child == null) {
+      return null;
+    }
+    return child.getSchema();
   }
 
   @Override
-  public int numTuples() {
-    if (uniqueTuples == null) {
-      return 0;
-    }
-    return uniqueTuples.numTuples();
-  }
-
-  /**
-   * Traverse through the list of tuples.
-   * */
-  private transient CheckUniquenessProcedure checkUniqueness;
-
-  /**
-   * Traverse through the list of tuples with the same hash code.
-   * */
-  private final class CheckUniquenessProcedure implements IntProcedure {
-
-    /** serialization id. */
-    private static final long serialVersionUID = 1L;
-
-    /** row index of the tuple. */
-    private int row;
-
-    /** input TupleBatch. */
-    private TupleBatch inputTB;
-
-    /** if found a replacement. */
-    private boolean unique;
-
-    @Override
-    public void value(final int index) {
-      if (TupleUtils.tupleEquals(inputTB, row, uniqueTuples, index)) {
-        unique = false;
-      }
-    }
-  };
-
-  @Override
-  public StreamingState duplicate() {
-    return new DupElim();
+  protected void init(final ImmutableMap<String, Object> execEnvVars) throws DbException {
+    initialToFinalhashCodes = new LongLongHashMap();
+    collidingFinalHashCodes = new LongObjectHashMap<LongArrayList>();
   }
 }
